Refactor routers

+++ ts/routes/api/animation-handlers.ts
@@ -5,34 +5,27 @@ import { Request } from "express";
 import { readCollection } from "../../src/queries/collection.js";
 
 export const handleCreateAnimation = async function (req: Request): Promise<boolean> {
-  if (!req.session.user || req.session.user.rank < 2) return false;
-  const authorId = req.session.user.id;
-  const collectionId = req.body.collectionId as number;
-  const frames = req.body.frames as number[];
-  const frameInterval = req.body.frameInterval as number;
+  const user = req.session.user;
+  const { collectionId, frames, frameInterval } = req.body;
+  if (!user || user.rank < 2) return false;
   if (!isValidId(collectionId) || !isValidFrameInterval(frameInterval) || !Array.isArray(frames)) return false;
-  if (frames.some((frame) => typeof frame !== "number")) return false;
+  if (frames.some((frame) => typeof frame !== "number" || !isValidId(frame))) return false;
   const collection = await readCollection(collectionId);
-  if (!collection || authorId !== collection.authorId) return false;
+  if (!collection || user.id !== collection.authorId) return false;
   const collectionMapsIds = collection.maps.map((map) => map.id);
-  for (let frame of frames) {
-    if (!isValidId(frame)) return false;
-    if (!collectionMapsIds.includes(frame)) return false;
-  }
+  if (!frames.every((frame) => collectionMapsIds.includes(frame))) return false;
   const animationRecord = await createAnimation(collectionId);
   if (animationRecord === null) return false;
-  const isGifSaved = await saveAnimationGif(animationRecord.id, frames, frameInterval);
+  await saveAnimationGif(animationRecord.id, frames, frameInterval);
   return true;
 };
 
 export const handleDeleteAnimation = async function (req: Request): Promise<boolean> {
-  if (!req.session.user || req.session.user.rank < 1) return false;
-  const userId = req.session.user.id;
+  const user = req.session.user;
   const animationId = parseInt(req.params.id);
-  if (!isValidId(animationId)) return false;
+  if (!user || user.rank < 1 || !isValidId(animationId)) return false;
   const animation = await readAnimation(animationId);
-  if (!animation) return false;
-  if (animation.collection.authorId !== userId) return false;
+  if (!animation || animation.collection.authorId !== user.id) return false;
   const isDeleted = await deleteAnimation(animationId);
   return isDeleted;
 };
